# packages/be/src/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Fastify API",
        "description": "API documentation for Luna",
        "version": "3.0.0"
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in the format 'Bearer <token>'"
            }
        },
...
    },
    "paths": {
...
        "/api/v1/users/": {
            "get": {
                "summary": "Retrieves a list of users",
                "tags": ["Users"],
                "description": "Retrieve all available users in the system.",
                "responses": {
                    "200": {
                        "description": "List of users.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "role": {
                                                "type": "string",
                                                "enum": [
                                                    "admin",
                                                    "procurementOfficer",
                                                    "researcher"
                                                ]
                                            },
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "firstName": {
                                                "type": "string"
                                            },
                                            "lastName": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "lastLoginDate": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "createdAt": {
                                                "type": "string"
                                            },
                                            "updatedAt": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "role",
                                            "id",
                                            "username",
                                            "firstName",
                                            "lastName",
                                            "email",
                                            "createdAt",
                                            "updatedAt"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error - Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "default": 401
                                        },
                                        "code": {
                                            "type": "string",
                                            "default": "FST_JWT_NO_AUTHORIZATION_IN_HEADER"
                                        },
                                        "error": {
                                            "type": "string",
                                            "default": "Unauthorized"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "No Authorization was found in request.headers"
                                        }
                                    },
                                    "required": ["statusCode", "code", "error", "message"]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates new user",
                "tags": ["Users"],
                "description": "Create user in the system.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 50
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "password"
                                    },
                                    "confirmPassword": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "confirmPassword"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": ["admin", "researcher", "procurementOfficer"]
                                    },
                                    "createdAt": {
                                        "type": "string"
                                    },
                                    "updatedAt": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "username",
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "password",
                                    "confirmPassword",
                                    "role"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created user's details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "role": {
                                            "type": "string",
                                            "enum": ["admin", "procurementOfficer", "researcher"]
                                        },
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "username": {
                                            "type": "string"
                                        },
                                        "firstName": {
                                            "type": "string"
                                        },
                                        "lastName": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "lastLoginDate": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "createdAt": {
                                            "type": "string"
                                        },
                                        "updatedAt": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "role",
                                        "id",
                                        "username",
                                        "firstName",
                                        "lastName",
                                        "email",
                                        "createdAt",
                                        "updatedAt"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error - Bad request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "default": 400
                                        },
                                        "code": {
                                            "type": "string",
                                            "default": "FST_ERR_VALIDATION"
                                        },
                                        "error": {
                                            "type": "string",
                                            "default": "Bad Request"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "body/password Password must be at least 8 characters long."
                                        }
                                    },
                                    "required": ["statusCode", "code", "error", "message"]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error - Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "default": 401
                                        },
                                        "code": {
                                            "type": "string",
                                            "default": "FST_JWT_NO_AUTHORIZATION_IN_HEADER"
                                        },
                                        "error": {
                                            "type": "string",
                                            "default": "Unauthorized"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "No Authorization was found in request.headers"
                                        }
                                    },
                                    "required": ["statusCode", "code", "error", "message"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}